lxd/db/snapshots: Return InstanceSnapshot from GetInstanceSnapshotsWithName
Currently, we use the parent instance's profiles for Snapshots, so
return an InstanceSnapshot struct instead of an Instance struct, to make
it clear that the ID of the entity is a snapshotID and not an instanceID
for retrieving profiles.

snapshot.ToAPI will convert to an api.Instancestruct and fill in the
Profiles via the instanceID obtained.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>