lxd/db/generate/db/method/v2: Expect collections for generated reference methods
There was an inconsistency in generated methods for tables that hold
reference data for other database entities (Devices, Config, Instance
Profiles, Certificate Projects) in that Update and Delete methods
operated on collections and Create expected a single entity.

This commit updates the generation for these methods to always expect a
collection. A collection is more ideal than single entities because
these sorts of relationships are often many-to-one relationships
(Devices to a Profile).

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>