lxd/db/generate/db/method/v2: Support multiple filter structs
Rather than replacing the internal fields of the filter structs with
slices, this commit changes generated GetMany methods to take multiple
filter structs. The first filter struct will be used as a baseline to
determine the prepared statement. Only matching fields will be
considered in the remaining filter structs. This is a bit of an awkward
implementation compared to using slices, but is far less intrusive to
the rest of the codebase.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>